requirements.txt
opencv-python
pygame
ultralytics
python-telegram-bot==20.0b1
python-dotenv
numpy

